SafeTask and Starter added, modified other code to work with them
SafeTask
========
Superclass of all 'Tasks, allows for safe Thread execution - gateway
variables so that threads don't have to be killed, which is dangerous
and throws exceptions.

Starter
=======
Is the task that starts comms (NetworkTable), and inititializes the Main
object main, WHICH NO LONGER HAS COMMS SIDE EFFECTS THAT CAUSE RUN ON
STARTUP CRASHES.

Includes camera initialization code.

importantDone() tells if we can continue with main and normal execution.