Bugfix - Teleop not starting
We hunted down a bug that was causing subsystems to not stop when told,
and would not allow the movement from disabledPeriodic to teleopInit to
teleopPeriodic.

Returning constants no longer starts threads. This was probably the
cause of the issue. Additionally, run() in safeTask is no longer
synchronized, as that prevent any kind of control from occurring.