Auto Debugging and Vision Tracking
Auto seems to work, although upon a second run without restarting the
code, it fails to complete some tasks. I need to find these pieces of
state that are not being reset and fix them.

Vision tracking now kind of has a hold on its fate, as when the errors
are close enough, it tells the loop to stop.

Spy-bot auto now working.